+++ addon/templates/components/frost-detail-tabs-more.hbs
@@ -1,6 +1,6 @@
 {{! Template for the qp-tabs-more component }}
 
-{{#network-element-details-v2/-components/qp-tab
+{{#frost-detail-tab
   class=(concat css '-tab')
   hook=(concat hookPrefix '-tab')
   register=(action '_register')
@@ -16,12 +16,12 @@
     pack='frost'
   }}
   {{_tab.label}}
-{{/network-element-details-v2/-components/qp-tab}}
+{{/frost-detail-tab}}
 
 {{#if (eq selectedTab.id _tab.id)}}
   {{to-elsewhere
     named=targetOutlet
-    send=(component 'network-element-details-v2/-components/qp-tabs-more-tabs'
+    send=(component 'frost-detail-tabs-more-tabs'
       tabs=_tabs
       onSelect=onSelect
     )
